requirements.txt
ajax==3.0.0
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==41.0.7
decorator==5.1.1
defusedxml==0.7.1
Django==4.2.7
django-allauth==0.58.2
django-appconf==1.0.6
django-filter==23.4
django-ranged-response==0.2.0
django-recaptcha==4.0.0
django-redis==5.4.0
django-simple-captcha==0.5.20
djangorestframework==3.14.0
drf-yasg==1.21.7
idna==3.6
inflection==0.5.1
kombu==5.3.4
oauthlib==3.2.2
packaging==23.2
Pillow==10.1.0
prompt-toolkit==3.0.41
psycopg2==2.9.6
psycopg2-binary==2.9.9
psycopg2-pool==1.1
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.12
